html/parsing/html5lib/html-tree/namespace-sensitivity.dat

#data
<body><table><tr><td><svg><td><foreignObject><span></td>Foo
#errors
no DOCTYPE
no </span>
text in table
EOF
#document
| <html>
|   <head>
|   <body>
|     "Foo"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 <svg foreignObject>
|                   <span>
